move preview to the left, remove preview panel, fix navigation buttons layout, prepare placeholder for entry details metadata
