+++ src/applications/content-entries-app/entry/entry-preview/entry-preview.component.html
@@ -1,70 +1,3 @@
-<div class="kEntryPreview kOverrideFAIcons">
+<div class="kEntryPreview">
 	<iframe *ngIf="_handler.iframeSrc" [src]="_handler.iframeSrc | safe" width="280" height="158" allowfullscreen webkitallowfullscreen mozAllowFullScreen frameborder="0"></iframe>
-	<div class="kPreviewAndEmbedContainer" *ngIf="_entryHasContent">
-		<i (click)="openPreviewAndEmbed()" [class.disabled]="!_entryReady" class="kIconshare"></i><a [class.disabled]="!_entryReady" (click)="openPreviewAndEmbed()">{{'applications.content.entryDetails.preview.pande' | translate}}</a>
-	</div>
-	<div class="kPreviewDetails" *ngIf="_handler.data">
-		<span class="kTitle">{{'applications.content.entryDetails.preview.details' | translate}}</span>
-
-		<div class="kRow">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.entryID' | translate}}</span>
-			<span>{{_handler.data.id }}</span>
-		</div>
-		<div class="kRow">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.creator' | translate}}</span>
-			<span [kTooltip]="_handler.data.creatorId" [placement]="'left'">{{_handler.data.creatorId}}</span>
-		</div>
-		<div class="kRow">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.creationDate' | translate}}</span>
-			<span>{{_handler.data.createdAt | date:'MM/dd/y HH:MM'}}</span>
-		</div>
-		<div class="kRow">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.updateDate' | translate}}</span>
-			<span>{{_handler.data.updatedAt | date:'MM/dd/y HH:MM'}}</span>
-		</div>
-		<div class="kRow"  *ngIf="_handler._landingPage">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.landingPage' | translate}}</span>
-			<a  [kTooltip]="_handler._landingPage" [placement]="'left'" [maxWidth]="150" (click)="openLandingPage(_handler._landingPage)">{{_handler._landingPage}}</a>
-		</div>
-		<div class="kRow" *ngIf="_isLive && _handler.data.recordedEntryId">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.vodEntry' | translate}}</span>
-			<a (click)="navigateToEntry(currentEntry.recordedEntryId)">{{_handler.data.recordedEntryId}}</a>
-		</div>
-		<div class="kRow" *ngIf="_isRecordedLive && _handler.data.rootEntryId">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.archiveOf' | translate}}</span>
-			<a (click)="navigateToEntry(currentEntry.rootEntryId)">{{_handler.data.rootEntryId}}</a>
-		</div>
-		<div class="kRow">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.type' | translate}}</span>
-			<span>{{_handler.data.mediaType | entryType: true}}</span>
-		</div>
-		<div class="kRow">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.moderation' | translate}}</span>
-			<span>{{_handler.data.moderationStatus | moderation}}</span>
-		</div>
-		<div class="kRow" *ngIf="_entryReady">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.plays' | translate}}</span>
-			<span>{{_handler.data.plays}}</span>
-		</div>
-		<div class="kRow" *ngIf="_entryReady && _handler.data.rank">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.rating' | translate}}</span>
-			<span>{{_handler.data.rank}}</span>
-		</div>
-		<div class="kRow" *ngIf="_hasDuration">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.duration' | translate}}</span>
-			<span>{{_handler.data.duration | kTime}}</span>
-		</div>
-		<div class="kRow" *ngIf="_entryReady && _handler.data.moderationCount">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.flags' | translate}}</span>
-			<span>{{_handler.data.moderationCount}}</span>
-		</div>
-		<div class="kRow" *ngIf="_entryReady && _handler.data.votes">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.votes' | translate}}</span>
-			<span>{{_handler.data.votes}}</span>
-		</div>
-		<div class="kRow" *ngIf="_isClip">
-			<span class="kLabels">{{'applications.content.entryDetails.preview.clippedFrom' | translate}}</span>
-			<span><a (click)="navigateToEntry(currentEntry.rootEntryId)">{{_handler.data.rootEntryId}}</a></span>
-		</div>
-	</div>
 </div>
