+++ src/applications/content-entries-app/entry/entry-sections-list/entry-sections-list.component.html
@@ -1,7 +1,9 @@
 <k-area-blocker [showLoader]="_loading">
-<ul *ngIf="_showList">
-	<li *ngFor="let section of _sections" [class.active]="section.attached">
-		<div [class.kSectionHasErrors]="!section.isValid" (click)="navigateToSection(section)">{{section.label}}</div>
-	</li>
-</ul>
+	<div class="kSections">
+		<ul *ngIf="_showList">
+			<li *ngFor="let section of _sections" [class.active]="section.attached">
+				<div [class.kSectionHasErrors]="!section.isValid" (click)="navigateToSection(section)">{{section.label}}</div>
+			</li>
+		</ul>
+	</div>
 </k-area-blocker>
