+++ src/applications/content-entries-app/entry/entry-sections-list/entry-sections-list.component.scss
@@ -1,5 +1,14 @@
 @import 'app-theme/_variables.scss';
 
+.kSections{
+	width: 100%;
+	height: 100%;
+	overflow-y: auto;
+	background-color: #ffffff;
+	border: 1px solid #ffffff;
+	border-radius: 3px;
+}
+
 ul {
 	list-style-type: none;
 	padding-left: 0px;
